requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
ffmpeg-python==0.2.0
frozenlist==1.3.1
future==0.18.2
idna==3.3
multidict==6.0.2
requests==2.28.1
six==1.16.0
urllib3==1.26.11
websocket-client==0.54.0
yarl==1.8.1
